refactor/dev-7142-world-map-to-typescript
