+++ packages/map/src/components/WorldMap/WorldMap.tsx
@@ -1,22 +1,21 @@
 import { memo, useContext } from 'react'
 
-import { jsx } from '@emotion/react'
 import ErrorBoundary from '@cdc/core/components/ErrorBoundary'
 import { geoMercator } from 'd3-geo'
 import { Mercator } from '@visx/geo'
 import { feature } from 'topojson-client'
 import topoJSON from './../data/world-topo.json'
-import ZoomableGroup from '../../ZoomableGroup'
-import Geo from '../../Geo'
-import CityList from '../../CityList'
-import BubbleList from '../../BubbleList'
-import ConfigContext from '../../../context'
+import ZoomableGroup from '../ZoomableGroup'
+import Geo from '../Geo'
+import CityList from '../CityList'
+import BubbleList from '../BubbleList'
+import ConfigContext from '../../context'
 
 const { features: world } = feature(topoJSON, topoJSON.objects.countries)
 
 let projection = geoMercator()
 
-const WorldMap = props => {
+const WorldMap = () => {
   // prettier-ignore
   const {
     applyLegendToRow,
@@ -127,7 +126,7 @@ const WorldMap = props => {
 
       const geoStrokeColor = state.general.geoBorderColor === 'darkGray' ? 'rgba(0, 0, 0, 0.2)' : 'rgba(255,255,255,0.7)'
 
-      let styles = {
+      let styles: Record<string, string | Record<string, string>> = {
         fill: '#E6E6E6',
         cursor: 'default'
       }
